fix(opamp): unify ServerToAgent builders across hot and push paths
The cross-server push path (ServerService.buildServerToAgentMessage) was a
self-described "simplified" stub that omitted remote_config, packages,
connection_settings and command. For a normal, fully-described agent it also
left the ReportFullState flag unset, so an immediate push to an agent connected
to another server delivered an effectively empty ServerToAgent; the config only
converged on the agent's next heartbeat via the OpAMP hot path.

Extract the complete message construction into a shared domain
ServerToAgentBuilder and route both the hot path (opamp.Service.fetchServerToAgent)
and the cross-server push path (ServerService.sendServerToAgentForInstance)
through it, so the two cannot diverge again. The unified builder is a strict
superset: it keeps the hot path's remote-config / packages / connection-settings /
command / server-capabilities and adds the ReportFullState flag when the agent is
not yet fully described.

Adds unit tests asserting the builder advertises the server capabilities and
delivers an assigned remote config.

Refs #501.
