+++ pkg/apiserver/domain/agent/service/server_test.go
@@ -367,6 +367,7 @@ func TestServerService_GetServer_CacheHit(t *testing.T) {
 		mockConnection,
 		mockAgent,
 		noopAgentCacheInvalidator{},
+		agentservice.NewServerToAgentBuilder(nil, slog.Default()),
 	)
 
 	fakeClock := newTestFakeClock(now)
@@ -414,6 +415,7 @@ func TestServerService_GetServer_CacheMiss_Expired(t *testing.T) {
 		mockConnection,
 		mockAgent,
 		noopAgentCacheInvalidator{},
+		agentservice.NewServerToAgentBuilder(nil, slog.Default()),
 	)
 
 	fakeClock := newTestFakeClock(now)
@@ -455,6 +457,7 @@ func TestServerService_GetServer_DatabaseError(t *testing.T) {
 		mockConnection,
 		mockAgent,
 		noopAgentCacheInvalidator{},
+		agentservice.NewServerToAgentBuilder(nil, slog.Default()),
 	)
 
 	_, err := svc.GetServer(ctx, serverID)
@@ -500,6 +503,7 @@ func TestServerService_GetServer_CacheUpdate(t *testing.T) {
 		mockConnection,
 		mockAgent,
 		noopAgentCacheInvalidator{},
+		agentservice.NewServerToAgentBuilder(nil, slog.Default()),
 	)
 
 	fakeClock := newTestFakeClock(now)
@@ -553,6 +557,7 @@ func TestServerService_SendMessageToServer_LocalShortCircuit(t *testing.T) {
 		mockConnection,
 		mockAgent,
 		noopAgentCacheInvalidator{},
+		agentservice.NewServerToAgentBuilder(nil, slog.Default()),
 	)
 	svc.SetClock(newTestFakeClock(now))
 
@@ -603,6 +608,7 @@ func TestServerService_SendMessageToServer_RemoteDispatch(t *testing.T) {
 		mockConnection,
 		mockAgent,
 		noopAgentCacheInvalidator{},
+		agentservice.NewServerToAgentBuilder(nil, slog.Default()),
 	)
 	svc.SetClock(newTestFakeClock(now))
 
@@ -655,6 +661,7 @@ func newLeaderTestService(
 		new(MockConnectionUsecase),
 		new(MockAgentUsecase),
 		noopAgentCacheInvalidator{},
+		agentservice.NewServerToAgentBuilder(nil, slog.Default()),
 	)
 	svc.SetClock(newTestFakeClock(now))
 
@@ -677,6 +684,7 @@ func newServerServiceForInvalidation(
 		new(MockConnectionUsecase),
 		new(MockAgentUsecase),
 		invalidator,
+		agentservice.NewServerToAgentBuilder(nil, slog.Default()),
 	)
 	svc.SetClock(newTestFakeClock(now))
